test: refactor large openapi test into smaller files
